chore: Split out `util` module into own file for `oneil_snapshot_tests`
